requirements.txt
gallery-dl==1.21.1
marshmallow==3.15.0
pika==1.2.0
pymongo==4.1.0
python-dotenv==0.20.0
redis==4.2.2